pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "marianalyzer"
version = "0.1.0"
description = "Local RAG system for document analysis"
authors = [{name = "Maxence Beaumont", email = "beaumont.maxence@example.com"}]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "typer[all]>=0.9.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "chromadb>=0.4.0",
    "rank-bm25>=0.2.2",
    "pypdf>=3.17.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.2",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "nltk>=3.8.1",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
]

[project.scripts]
marianalyzer = "marianalyzer.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["marianalyzer*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
